# Use cublasSgemmBatched when LHS is minibatched.