This pull request replaces multiple `cublasSgemm` call to `cublasSgemmBatched` and makes the speed of `functions::matmul` faster when the left-hand side is minbatched.

Benchmark code (C++):

```c++
#include <primitiv/primitiv.h>
using namespace primitiv;
int main() {
  devices::CUDA dev(0); Device::set_default(dev);
  Graph g; Graph::set_default(g);
  for (int i = 0; i < 1024; ++i) {
    g.clear();
    Node lhs = functions::ones<Node>(Shape({256, 256}, L));
    Node rhs = functions::ones<Node>(Shape({256, 256}, R));
    Node result = functions::matmul(lhs, rhs);
    result.backward();
  }
  return 0;
}
```

Results on GeForce GTX1080:

```
# L == 1, R == 1024 (no effect by this branch)
before: 16.79s user 7.56s system 99% cpu 24.556 total
after:  15.90s user 8.38s system 99% cpu 24.443 total

# L == 1024, R == 1
before: 27.42s user 7.41s system 99% cpu 34.978 total
after:  15.90s user 6.78s system 99% cpu 22.831 total

# L == 1024, R == 1024
before: 30.03s user 8.78s system 99% cpu 38.957 total
after:  16.65s user 8.30s system 99% cpu 25.103 total
```